+++ lib/MusicMachine/Instrument.h
@@ -37,7 +37,7 @@ namespace HLMusicMachine
     CircularBuffer<NoteEvent, 256> eventsBuffer;
 
     std::vector<Voice *> voices;
-    int voiceIndex = 0;
+    uint8_t voiceIndex = 0;
 
     const int minTrigTime = 1 * TASK_MILLISECOND;
     const int maxTrigTime = 100 * TASK_MILLISECOND;
@@ -45,10 +45,10 @@ namespace HLMusicMachine
     bool isEnabled = false;
     void toggleEnabled();
 
-    void trigNote(int note = 0, int vel = 127, int noteLenght = 0);
+    void trigNote(uint8_t note = 0, uint8_t vel = 127, uint16_t noteLenght = 0);
 
-    virtual void noteOn(int note = 0, int vel = 127, uint8_t voiceIndex = 0);
-    virtual void noteOff(int note = 0, uint8_t voiceIndex = 0);
+    virtual void noteOn(uint8_t note = 0, uint8_t vel = 127, uint8_t voiceIndex = 0);
+    virtual void noteOff(uint8_t note = 0, uint8_t voiceIndex = 0);
     virtual void randomize();
     virtual void pitchBend(double val);
